requirements.txt
PyYAML~=6.0.1
pprintpp~=0.4.0
icecream~=2.1.3
prettytable~=3.9.0
